clean up thumbnail file
